Refactor out business logic and database functions from routes
